Fixed buggy period index conversion

+++ androidweekview/src/main/java/com/alamkanak/weekview/loaders/WeekViewLoader.kt
@@ -16,7 +16,7 @@ interface WeekViewLoader<T> {
 	 * @param instance the date
 	 * @return The period index in which the date falls (floating point number).
 	 */
-	fun toWeekViewPeriodIndex(instance: Calendar): Double
+	fun toWeekViewPeriodIndex(instance: Calendar): Long
 
 	/**
 	 * Load the events within the period
